Create VolumeCalcUtil to adjust volume of sound, tie into line control
